fixed bug in get_Eps returning int and not double

+++ include/signal.h
@@ -153,7 +153,7 @@ public:
 		Eps=val;
 	}
 
-	inline static int get_Eps(){
+	inline static double get_Eps(){
 		return Eps;
 	}
